Fix to fast_retrieve, refactored retrieve_controllers into local and master script files Fix to update_fast_load to return status
