+++ apis_v1/views/views_retrieve_tables.py
@@ -7,8 +7,9 @@ from django.http import HttpResponse
 
 import wevote_functions.admin
 from config.base import get_environment_variable
-from retrieve_tables.controllers import retrieve_sql_tables_as_csv, fast_load_status_retrieve, get_total_row_count, \
-    fast_load_status_update
+from retrieve_tables.controllers_master import fast_load_status_retrieve, retrieve_sql_tables_as_csv, \
+    get_total_row_count
+from retrieve_tables.controllers_master import fast_load_status_update
 from wevote_functions.functions import get_voter_device_id
 
 logger = wevote_functions.admin.get_logger(__name__)
