+++ retrieve_tables/urls.py
@@ -4,12 +4,11 @@
 
 from django.urls import re_path
 
-from . import controllers
+from retrieve_tables.controllers_local import retrieve_sql_files_from_master_server
+from retrieve_tables.controllers_master import fast_load_status_retrieve
 
 urlpatterns = [
     # views_admin
-    re_path(r'^import/status/$',
-        controllers.fast_load_status_retrieve, name='fast_load_status_retrieve'),
-    re_path(r'^import/files/$',
-        controllers.retrieve_sql_files_from_master_server, name='retrieve_sql_files_from_master_server'),
+    re_path(r'^import/status/$', fast_load_status_retrieve, name='fast_load_status_retrieve'),
+    re_path(r'^import/files/$', retrieve_sql_files_from_master_server, name='retrieve_sql_files_from_master_server'),
 ]
